Added way to handle precision - now the program handles precision upto two decimal place - fixed a couple of small things

+++ 01-Banking-System/Account.h
@@ -20,9 +20,9 @@ class Account{
     public:
         Account(long long amount);
 
-        void deposit(long long amount);
+        void deposit(double amount);
 
-        bool withdraw(long long amount);
+        bool withdraw(double amount);
 
         long long getBalance() const;
